+++ src/main/org/jboss/messaging/core/remoting/impl/wireformat/SessionBrowserHasNextMessageMessage.java
@@ -36,13 +36,13 @@ public class SessionBrowserHasNextMessageMessage extends PacketImpl
 
    // Attributes ----------------------------------------------------
 
-   private int browserID;
+   private long browserID;
 
    // Static --------------------------------------------------------
 
    // Constructors --------------------------------------------------
 
-   public SessionBrowserHasNextMessageMessage(final int browserID)
+   public SessionBrowserHasNextMessageMessage(final long browserID)
    {
       super(SESS_BROWSER_HASNEXTMESSAGE);
 
@@ -56,19 +56,19 @@ public class SessionBrowserHasNextMessageMessage extends PacketImpl
 
    // Public --------------------------------------------------------
 
-   public int getBrowserID()
+   public long getBrowserID()
    {
       return browserID;
    }
    
    public void encodeBody(final MessagingBuffer buffer)
    {
-      buffer.putInt(browserID);
+      buffer.putLong(browserID);
    }
    
    public void decodeBody(final MessagingBuffer buffer)
    {
-      browserID = buffer.getInt();       
+      browserID = buffer.getLong();       
    }
 
    @Override
